# gorestful/apidocs.json
{
 "swagger": "2.0",
 "info": {
  "description": "Resource for managing Users",
  "title": "UserService",
  "contact": {
   "name": "john",
   "url": "http://johndoe.org",
   "email": "john@doe.rp"
  },
  "license": {
   "name": "MIT",
   "url": "http://mit.org"
  },
  "version": "1.0.0"
 },
 "paths": {
  "/users": {
   "get": {
    "consumes": [
     "application/xml",
     "application/json"
    ],
    "produces": [
     "application/json",
     "application/xml"
    ],
    "tags": [
     "users"
    ],
    "summary": "get all users",
    "operationId": "findAllUsers",
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "type": "array",
       "items": {
        "$ref": "#/definitions/main.User"
       }
      }
     }
    }
   },
   "post": {
    "consumes": [
     "application/xml",
     "application/json"
    ],
    "produces": [
     "application/json",
     "application/xml"
    ],
    "tags": [
     "users"
    ],
    "summary": "create a user",
    "operationId": "createUser",
    "parameters": [
     {
      "name": "body",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/main.User"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK"
     }
    }
   }
  },
  "/users/{user-id}": {
   "get": {
    "consumes": [
     "application/xml",
     "application/json"
    ],
    "produces": [
     "application/json",
     "application/xml"
    ],
    "tags": [
     "users"
    ],
    "summary": "get a user",
    "operationId": "findUser",
    "parameters": [
     {
      "type": "integer",
      "default": 1,
      "description": "identifier of the user",
      "name": "user-id",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "schema": {
       "$ref": "#/definitions/main.User"
      }
     },
     "404": {
      "description": "Not Found"
     }
    }
   },
   "put": {
    "consumes": [
     "application/xml",
     "application/json"
    ],
    "produces": [
     "application/json",
     "application/xml"
    ],
    "tags": [
     "users"
    ],
    "summary": "update a user",
    "operationId": "updateUser",
    "parameters": [
     {
      "type": "string",
      "description": "identifier of the user",
      "name": "user-id",
      "in": "path",
      "required": true
     },
     {
      "name": "body",
      "in": "body",
      "required": true,
      "schema": {
       "$ref": "#/definitions/main.User"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK"
     }
    }
   },
   "delete": {
    "consumes": [
     "application/xml",
     "application/json"
    ],
    "produces": [
     "application/json",
     "application/xml"
    ],
    "tags": [
     "users"
    ],
    "summary": "delete a user",
    "operationId": "removeUser",
    "parameters": [
     {
      "type": "string",
      "description": "identifier of the user",
      "name": "user-id",
      "in": "path",
      "required": true
     }
    ],
    "responses": {
     "200": {
      "description": "OK"
     }
    }
   }
  }
 },
 "definitions": {
  "main.User": {
   "required": [
    "id",
    "name",
    "age"
   ],
   "properties": {
    "age": {
     "description": "age of the user",
     "type": "integer",
     "format": "int32",
     "default": 21
    },
    "id": {
     "description": "identifier of the user",
     "type": "string"
    },
    "name": {
     "description": "name of the user",
     "type": "string",
     "default": "john"
    }
   }
  }
 },
 "tags": [
  {
   "description": "Managing users",
   "name": "users"
  }
 ]
}